+++ lib/lock.sh
@@ -6,16 +6,21 @@
 trace "Locking process"
 
 # Define lock file location
-LOCK_FILE=/tmp/$APP.lock
+LOCK_FILE="/tmp/$APP.lock"
 
 function lock() {
-  trace "Creating lockfile"
 
-  if [ -f "$LOCK_FILE" ]; then
-    warning "${WORKPATH}/${APP} is already being deployed in another instance."
-    exit
-  fi
-
-  trap 'rm -f "${LOCK_FILE}"' EXIT
-  touch "${LOCK_FILE}"
+	if [[ -f "${LOCK_FILE}" ]]; then
+		# Unlock?
+		if [[ "${UNLOCK}" == "1" ]]; then
+			rm "${LOCK_FILE}"
+			notice "${WORKPATH}/${APP} is now unlocked."; quietExit
+		else
+			warning "${WORKPATH}/${APP} is already being deployed in another instance."; quietExit
+		fi
+	else
+		trap 'rm -f "${LOCK_FILE}"' EXIT
+		trace "Creating lockfile"
+  		touch "${LOCK_FILE}"
+	fi
 }