+++ CHANGELOG.md
@@ -11,6 +11,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 - Added integration with Google Analytics for web projects
 - Brought back client emails in the form of a weekly digest
 - Added Wordpress database checks
+- Added `deploy --unlock` to make it easier remove stale .lock files
 - Added options to define custom file paths used in your Wordpress projects
 - Added email and Slack notification for automated deployments when no updates are available
 - Missing client logos no longer break HTML emails and logs